Refactored RecentChips out of RecentChipsContainer
